# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 小确幸服务端 API
  version: 0.1.0
  description: |
    S0–S3 surface: health, auth, personal space, records, sync, photo STS upload.
    Field names are lowerCamelCase. Timestamps are UTC RFC3339.
    Server IDs are UUIDv7 strings.
    Media quota: 200MB/user, 5MB/photo, 9 photos/record. Invite: link only, 7 days, max 6 members. Account delete cooling: 24h.
servers:
  - url: /
paths:
...
  /api/v1/auth/sms/send:
    post:
      operationId: smsSend
      summary: Send a login SMS. Response does not reveal whether the phone is registered.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/DeviceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, examples: ["13812345678"] }
                deviceId: { type: string }
      responses:
        "200":
          description: Accepted (always, to avoid account enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /api/v1/auth/sms/verify:
    post:
      operationId: smsVerify
      summary: Verify SMS code and issue access/refresh tokens. Creates the user on first success.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/DeviceID"
        - $ref: "#/components/parameters/Platform"
        - $ref: "#/components/parameters/AppVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
                deviceId: { type: string }
                platform: { type: string }
                appVersion: { type: string }
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/token/refresh:
    post:
      operationId: tokenRefresh
      summary: Rotate refresh token. Reuse of a consumed token revokes the family.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
                deviceId: { type: string }
      responses:
        "200":
          description: New pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Revoke the refresh-token family
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Logged out
  /api/v1/account:
    delete:
      operationId: deleteAccount
      summary: Mark the account PENDING_DELETE and revoke sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending delete
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me:
    get:
      operationId: getMe
      security: [{ bearerAuth: [] }]
      summary: Current account. Phone is masked.
      responses:
        "200": { description: Profile }
    patch:
      operationId: patchMe
      security: [{ bearerAuth: [] }]
      summary: Update display name (1-20 chars)
      responses:
        "200": { description: Profile }
  /api/v1/stats/calendar:
    get:
      operationId: moodCalendar
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: spaceId, in: query, required: true, schema: { type: string } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses:
        "200": { description: Daily mood and GP }
  /api/v1/spaces:
    get:
      operationId: listSpaces
      security: [{ bearerAuth: [] }]
      summary: List spaces the current user can access
      responses:
        "200":
          description: Spaces
  /api/v1/spaces/{id}:
    get:
      operationId: getSpace
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Space }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/spaces/{id}/plant:
    get:
      operationId: getPlant
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Plant derived from space GP }
  /api/v1/records:
    get:
      operationId: listRecords
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: spaceId, in: query, required: true, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200": { description: Records }
    post:
      operationId: createRecord
      security: [{ bearerAuth: [] }]
      summary: Online create; same GP path as sync push
      responses:
        "200": { description: Applied mutation }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/records/{id}:
    get:
      operationId: getRecord
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Record }
    patch:
      operationId: patchRecord
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Applied }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteRecord
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Tombstone }
  /api/v1/sync/push:
    post:
      operationId: syncPush
      security: [{ bearerAuth: [] }]
      summary: Offline mutation batch. Per-item transactions.
      responses:
        "200": { description: Per-mutation results }
  /api/v1/sync/pull:
    get:
      operationId: syncPull
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200": { description: Incremental changes including tombstones }
  /api/v1/media/sts:
    post:
      operationId: mediaSTS
      security: [{ bearerAuth: [] }]
      summary: Issue a short-lived PUT URL. Android uploads directly; server never sees a long-lived AK on the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mimeType, sizeBytes]
              properties:
                type: { type: string, enum: [PHOTO, VOICE] }
                mimeType: { type: string, examples: ["image/jpeg"] }
                sizeBytes: { type: integer }
                width: { type: integer }
                height: { type: integer }
      responses:
        "200": { description: mediaId, uploadUrl, headers, quota }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/media/complete:
    post:
      operationId: mediaComplete
      security: [{ bearerAuth: [] }]
      summary: Verify the object landed and mark READY.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mediaId]
              properties:
                mediaId: { type: string }
      responses:
        "200": { description: READY object }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/media/quota:
    get:
      operationId: mediaQuota
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: usedBytes / maxBytes (200MB) }
  /api/v1/media/{id}/download-url:
    get:
      operationId: mediaDownloadURL
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Short-lived GET URL }
  /api/v1/media/{id}:
    delete:
      operationId: mediaDelete
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Soft-deleted }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/media/upload/{ticket}:
    put:
      operationId: mediaMockUpload
      summary: Mock/OSS-compat PUT target when OSS_PROVIDER=mock
      parameters:
        - { name: ticket, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Stored }
  /api/v1/media/content/{ticket}:
    get:
      operationId: mediaMockDownload
      summary: Mock signed GET when OSS_PROVIDER=mock
      parameters:
        - { name: ticket, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Bytes }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      schema: { type: string }
    DeviceID:
      name: X-Device-ID
      in: header
      schema: { type: string }
    Platform:
      name: X-Platform
      in: header
      schema: { type: string, examples: ["android"] }
    AppVersion:
      name: X-App-Version
      in: header
      schema: { type: string }
  schemas:
    Meta:
      type: object
      properties:
        requestId: { type: string }
        serverTime: { type: string, format: date-time }
    Success:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/Meta" }
    Error:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
            retryable: { type: boolean }
        meta: { $ref: "#/components/schemas/Meta" }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"